Moved Transpose LibraryNode to STD
