+++ dace/libraries/blas/nodes/__init__.py
@@ -1,11 +1,10 @@
-# Copyright 2019-2021 ETH Zurich and the DaCe authors. All rights reserved.
+# Copyright 2019-2023 ETH Zurich and the DaCe authors. All rights reserved.
 from .matmul import MatMul
 from .dot import Dot
 from .gemv import Gemv
 from .gemm import Gemm
 from .ger import Ger
 from .batched_matmul import BatchedMatMul
-from .transpose import Transpose
 
 from .axpy import Axpy
 from .einsum import Einsum
