+++ dace/libraries/linalg/nodes/solve.py
@@ -1,13 +1,14 @@
-# Copyright 2019-2021 ETH Zurich and the DaCe authors. All rights reserved.
+# Copyright 2019-2023 ETH Zurich and the DaCe authors. All rights reserved.
 import copy
 import dace
 import dace.library
 import dace.properties
 import dace.sdfg.nodes
 import numpy as np
+
 from dace import Memlet
-from dace.libraries.blas.nodes import Transpose
-from dace.libraries.lapack.nodes import Getrf, Getrs
+from dace.libraries.lapack import Getrf, Getrs
+from dace.libraries.standard import Transpose
 from dace.transformation.transformation import ExpandTransformation
 from dace.libraries.lapack import environments
 from dace.libraries.blas import environments as blas_environments
